MIDI-32: More modular training loop
