+++ configs/config-single-batch.yaml
@@ -3,6 +3,7 @@ hydra:
     chdir: False
 
 train:
+  task: "binary-classification"
   dataset_name: "roszcz/ecg-segmentation-ltafdb" # huggingface dataset
   batch_size: 16
   num_workers: 4
